requirements.txt
google-genai>=0.6.0
python-dotenv>=1.0.1
pyttsx3>=2.90
SpeechRecognition>=3.10.4
